backend/requirements.txt
fastapi>=0.104.1
openai>=1.3.5
cohere>=5.5.3
qdrant-client>=1.8.0
python-dotenv>=1.0.0
uvicorn>=0.24.0
pydantic>=2.5.0
requests>=2.31.0
trafilatura>=1.12.2